OHRM5X-1146: Requested changes to admin access
