+++ symfony/plugins/orangehrmAuthenticationPlugin/Subscriber/AdministratorAccessSubscriber.php
@@ -35,7 +35,7 @@ class AdministratorAccessSubscriber extends AbstractEventSubscriber
     /**
      * @inheritDoc
      */
-    public static function getSubscribedEvents()
+    public static function getSubscribedEvents(): array
     {
         return [
             KernelEvents::CONTROLLER => [
