+++ symfony/plugins/orangehrmAuthenticationPlugin/config/routes.yaml
@@ -13,12 +13,7 @@ auth_logout:
   controller: OrangeHRM\Authentication\Controller\LogoutController::handle
   methods: [ GET ]
 
-auth_admin_access:
-  path: /auth/adminAccess
-  controller: OrangeHRM\Authentication\Controller\AdministratorAccessController::handle
-  methods: [ GET ]
-
 auth_verify:
   path: /auth/adminVerify
-  controller: OrangeHRM\Authentication\Controller\VerifyController::handle
+  controller: OrangeHRM\Authentication\Controller\AdministratorVerifyController::handle
   methods: [ POST ]
